requirements.txt
opencv-python == 4.6.0.66
click == 8.1.3
requests == 2.28.1
pillow == 9.3.0